Refactor: Refactor functions to process dataSource to tools.tsx
