+++ apps/expo/src/screens/home/DeviceOverviewScreen/_deviceList.tsx
@@ -11,6 +11,7 @@ import useCloudFeeds from "@/hooks/cloud-feed/useCloudFeeds";
 import useDevices from "@/hooks/device/useDevices";
 import { UserContext } from "@/providers/UserProvider";
 import { BuildingDeviceResponse, DataSourcesList } from "@/types/api";
+import { processDataSource } from "@/utils/tools";
 
 export default function DeviceList({
   buildingId,
@@ -34,84 +35,20 @@ export default function DeviceList({
   const [allItemsDone, setAllItemsDone] = useState(true);
   const connectedState: number[] = [];
 
-  function checkStatus(
-    dataSource: {
-      id: number;
-      type: { name: string };
-      item: { id: number; name: string; installation_manual_url: string; info_url: string };
-      precedes: { id: number }[];
-      uploadschedule: string;
-    },
-    oldSource: BuildingDeviceResponse
-  ) {
-    const activated_at = oldSource?.activated_at ?? null;
-    if (
-      (dataSource.type.name === "cloud_feed" &&
-        cloudFeedData?.find(item => item.cloud_feed.name === dataSource.item.name)?.connected) ||
-      !(activated_at === null)
-    ) {
-      return 2;
-    }
-    return 1;
-  }
-
   useEffect(() => {
     if (dataSourcesList) {
       let connectedCount = 0;
       const newData: BuildingDeviceResponse[] = [];
 
       dataSourcesList.items.forEach(dataSource => {
-        let connectStatus = 1;
-        const oldSource = data?.find(item => item.device_type.name === dataSource.item.name);
-        const activated_at = oldSource?.activated_at ?? null;
-        const latest_upload = oldSource?.latest_upload ?? null;
-        const upload_schedule = dataSource.uploadschedule;
-
-        if (oldSource) {
-          connectStatus = checkStatus(dataSource, oldSource);
-          if (connectStatus === 2) connectedState.push(dataSource.id);
-        }
-
-        //Check if all precedes are completed
-        const itemsNotPrecedingCurrent = dataSourcesList.items.filter(otherItem => {
-          const precedesMatch = otherItem.precedes.some(precede => precede.id === dataSource.id);
-          return otherItem.id !== dataSource.id && precedesMatch;
-        });
-
-        let allPrecedesDone = true;
-        if (itemsNotPrecedingCurrent.length > 0) {
-          itemsNotPrecedingCurrent.forEach(otherItem => {
-            const otherOldSource = data?.find(item => item.device_type.name === otherItem.item.name);
-            if (otherOldSource) {
-              if (checkStatus(otherItem, otherOldSource) === 1) {
-                allPrecedesDone = false;
-              }
-            } else {
-              allPrecedesDone = false;
-            }
-          });
-        }
-
-        connectStatus = connectStatus === 2 ? connectStatus : allPrecedesDone ? 0 : 1;
+        const newResponse = processDataSource(dataSource, data, cloudFeedData, dataSourcesList, buildingId);
+        newData.push(newResponse);
 
         //Progressbar
-        if (connectStatus === 2) {
+        if (newResponse.connected === 2) {
+          connectedState.push(dataSource.id);
           connectedCount++;
         }
-
-        const newResponse: BuildingDeviceResponse = {
-          id: dataSource.id,
-          name: oldSource?.name ? oldSource.name : dataSource.item.name,
-          building_id: buildingId,
-          device_type: dataSource.item,
-          activated_at,
-          latest_upload,
-          upload_schedule,
-          typeCategory: dataSource.type.name,
-          connected: connectStatus,
-          notification_threshold_duration: dataSource.notificationThresholdDuration,
-        };
-        newData.push(newResponse);
       });
 
       setItemData(newData);
