+++ apps/expo/src/types/api.ts
@@ -94,21 +94,18 @@ const exampleItems = [
 //TODO: TEMPORARY HARDOCODE, AWAIT API UPDATE THEN REMOVE THE DEFAULTS, OPTIONAL AND NULLABLE
 export type DataSourcesList = {
   description: string;
-  items: {
-    id: number;
-    type: {
-      name: string;
-    };
-    item: {
-      id: number;
-      name: string;
-      installation_manual_url: string;
-      info_url: string;
-    };
-    precedes: { id: number }[];
-    uploadschedule: string;
-    notificationThresholdDuration: string;
-  }[];
+  items: DataSourceType[];
+};
+
+export type DataSourceType = {
+  id: number;
+  type: {
+    name: string;
+  };
+  item: DataSourceItemType;
+  precedes: { id: number }[];
+  uploadschedule: string;
+  notificationThresholdDuration: string;
 };
 
 export type DataSourceItemType = {
